+++ recipes/default.rb
@@ -1,21 +1,14 @@
-#
-# Author:: Guilhem Lettron (<guilhem.lettron@youscribe.com>)
-# Cookbook:: webpi
-# Recipe:: default
-#
-# Copyright:: 2011-2019, Chef Software, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
 
-include_recipe "webpi::install-#{node['webpi']['install_method']}"
+if node['webpi']['install_method'] == 'msi'
+  webpi_install 'msi' do
+    package_name node['webpi']['msi_package_name']
+    source node['webpi']['msi']
+    checksum node['webpi']['msi_checksum']
+  end
+else
+  webpi_install 'zip' do
+    source node['webpi']['url']
+    checksum node['webpi']['checksum']
+    install_directory node['webpi']['home']
+  end
+end