+++ resources/application.rb
@@ -1,36 +1,47 @@
-#
-# Author:: Madan Kapoor (<madankapoor10@gmail.com>)
-# Cookbook:: webpi
-# Resource:: application
-#
-# Copyright:: 2011-2019, Madan Kapoor
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
+property :app_id,
+        String,
+        name_property: true
 
-property :app_id, String, name_property: true
-property :accept_eula, [true, false], default: false
-property :suppress_reboot, [true, false], default: true
-property :iis_express, [true, false], default: false
-property :other_options, String # Can be used for Language or other options depending on application
-property :returns, [Integer, Array], default: [0, 42]
-property :sql_password, String, sensitive: true # To be used only if required
-property :mysql_password, String, sensitive: true # To be used only if required
-property :webpi_cmd_path, String, default: "#{ENV['SYSTEMDRIVE']}\\webpi\\WebpiCmd.exe"
-property :webpi_log_path, String, default: lazy { "#{Chef::Config[:file_cache_path]}/WebPI.log" }
-property :xml_path, String
+property :accept_eula,
+        [true, false],
+        default: false
 
-include Windows::Helper
+property :suppress_reboot,
+        [true, false],
+        default: true
+
+property :iis_express,
+        [true, false],
+        default: false
+
+property :other_options,
+        String,
+        description: 'Can be used for Language or other options depending on application'
+
+property :returns,
+        [Integer, Array],
+        default: [0, 42]
+
+property :sql_password,
+        String,
+        sensitive: true
+
+property :mysql_password,
+        String,
+        sensitive: true
+
+property :webpi_cmd_path,
+        String,
+        default: "#{ENV['SYSTEMDRIVE']}\\webpi\\WebpiCmd.exe"
+
+property :webpi_log_path,
+        String,
+        default: lazy { "#{Chef::Config[:file_cache_path]}/WebPI.log" }
+
+property :xml_path,
+        String
+
+unified_mode true
 
 action :install do
   install_list = apps_to_be_installed
@@ -62,6 +73,7 @@ action_class do
     cmd = "\"#{new_resource.webpi_cmd_path}\" /List /ListOption:Installed"
     cmd << " /XML:#{new_resource.xml_path}" if new_resource.xml_path
     cmd_out = shell_out(cmd, returns: [0, 42])
+
     if cmd_out.stderr.empty?
       new_resource.app_id.split(',').each do |p|
         # Example output
@@ -76,6 +88,7 @@ action_class do
       Chef::Log.info(cmd_out.stderr)
       install_array = new_resource.app_id
     end
+
     install_array.join(',')
   end
 end