+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # webpi Cookbook CHANGELOG
+
 This file is used to list changes made in each version of the webpi cookbook.
 
 ## Unreleased
 
-- resolved cookstyle error: recipes/install-msi.rb:29:21 warning: `Lint/SendWithMixinArgument`
+- Resolved cookstyle error: recipes/install-msi.rb:29:21 warning: `Lint/SendWithMixinArgument`
+- Add the webpi_install resource
 
 ## 4.2.1 (2017-05-30)
 
@@ -33,40 +35,45 @@ This file is used to list changes made in each version of the webpi cookbook.
 - Update the minimum chef requirement to 12.1 from 12.0
 
 ## 3.0.0 (2016-09-06)
+
 - Remove support for Chef 11
 
 ## 2.1.0 (2016-09-01)
+
 - Resolve rubocop warnings
 - Add use_inline_resources to the provider
 - Add chef_version metadata
 - Testing updates
 
 ## v2.0.0 (2016-01-04)
+
 - Install WebPI 5.0 by default
 - Fix runs on Chef 11 systems
 
 ## v1.2.9 (2015-12-30)
-### Improvement
+
 - Modernization of cookbook supporting travis ci, kitchen convergence testing, style checking and foodcritic rule compliance
 - Removed unnecessary includes of the default recipe in the windows cookbook to eliminate warnings
 
 ## v1.2.8 (2014-02-14)
-### Bug
+
 - **[COOK-3832](https://tickets.chef.io/browse/COOK-3832)** - install-msi recipe should leverage chef cache
 
 ## v1.2.6
+
 - [COOK-3372] - unzip webpi if webpi executable doesn't exist
 
 ## v1.2.4
-### Improvement
+
 - **[COOK-3474](https://tickets.chef.io/browse/COOK-3474)** - Upgrade WebPI (4.5 no longer exists)
 
 ## v1.2.2
-### Improvement
+
 - [COOK-2719]: update webpi documentation to show that Windows 2012
 - and Windows 8 are supported
 
 ## v1.2.0
+
 - [COOK-2298] - webpi will not install anything due to logic bug
 - [COOK-1270] - webpi creates logs with timestamps in the name
 - cluttering c:\
@@ -75,15 +82,19 @@ This file is used to list changes made in each version of the webpi cookbook.
 - components to determine which to install if any
 
 ## v1.1.0
+
 - [COOK-1251] - Fix LWRP "NotImplementedError"
 - [COOK-1726, COOK-1760] - Upgrade to webpi version 4
 
 ## v1.0.4
+
 - Created the ability to use the XML parameter for custom xml files.
 
 ## v1.0.2
+
 - Ruby 1.9 compat fixes
 - don't unzip if files already exist
 
 ## v1.0.0
+
 - initial release